third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header.html

<!DOCTYPE html>
<iframe src="resources/sandbox.php?sandbox=allow-scripts"></iframe>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.onload = function() {
    var frame = document.getElementsByTagName("iframe")[0];
    try {
        var secret = frame.contentWindow.secret;
        console.log("FAIL: Iframe was not in a unique origin: " + secret);
    } catch (e) {
        console.log("PASS: Iframe was in a unique origin");
    }
};
</script>
