third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header-expected.txt

CONSOLE MESSAGE: line 4: Script executed in iframe.
CONSOLE MESSAGE: line 13: PASS: Iframe was in a unique origin
